rename a folder and a file for clarity
